+++ src/about.js
@@ -10,7 +10,7 @@ const AboutPage = () =>
     <Row>
       <Col md={12}>
         <Jumbotron style={{ textAlign: 'center' }}>
-          <h1>Squid Tracks</h1>
+          <h1>SquidTracks</h1>
           <h2>A Splatnet 2 Client for your Desktop</h2>
           <h5>
             {`Beta Version ${appVersion} `}
@@ -27,17 +27,17 @@ const AboutPage = () =>
         </Jumbotron>
         <h2>Introduction</h2>
         <p>
-          Squid Tracks is a desktop program for tracking your Splatnet 2
+          SquidTracks is a desktop program for tracking your Splatnet 2
           statistics. It reads the data from the Nintendo Splatnet API and
           presents it in an alternative format that is hopefully more easy to
           digest. There is also some data that the Nintendo App does not show.
-          Squid Tracks offers that data up too.
+          SquidTracks offers that data up too.
         </p>
         <p>
-          Another feature of Squid Tracks is that it can back up your data to
-          stat.ink. Splatnet only records your most recent 50 games. Splat
-          Tracks offers integration with stat.ink where you can backup all of
-          your games and also offers them for public viewing on the web.
+          Another feature of SquidTracks is that it can back up your data to
+          stat.ink. Splatnet only records your most recent 50 games. SquidTracks
+          offers integration with stat.ink where you can backup all of your
+          games and also offers them for public viewing on the web.
         </p>
         <p>
           The records tab shows your win-loss records by stage and by weapon.
@@ -104,7 +104,6 @@ const AboutPage = () =>
             When done playing, click on the button again to stop auto uploads.
           </li>
         </ol>
-
         <h4>Manual Uploading</h4>
         <p>
           Automatic uploading only uploads new games. If there are older games
@@ -125,7 +124,14 @@ const AboutPage = () =>
             Click <strong>Upload to stat.ink</strong>
           </li>
         </ol>
-
+        <h2>Tracking and Privacy</h2>
+        This program uses google analytics to track version uptake, activity,
+        bugs, and crashing. Your information is anonymized before sending and
+        will not be shared with any third party. If you find this creepy you can
+        disable tracking in{' '}
+        <Link to="/settings">
+          <strong>Settings</strong>
+        </Link>.
         <h2>For Help and Bug Filing</h2>
         <p>Send cries for help, bug reports, and feature requests to</p>
         <p>
@@ -146,7 +152,6 @@ const AboutPage = () =>
             @SquidTracks
           </a>
         </p>
-
         <h2>Releases</h2>
         <p>
           Once the program is installed, the program should autoupdate on