+++ src/components/result-detail-card.js
@@ -1,5 +1,18 @@
 import React from 'react';
-import { Grid, Row, Col, Panel, Table, Glyphicon } from 'react-bootstrap';
+import {
+  Grid,
+  Row,
+  Col,
+  Panel,
+  Table,
+  Glyphicon,
+  ButtonToolbar,
+  ButtonGroup,
+  Button
+} from 'react-bootstrap';
+import TeamStatsTable from './team-stats-table';
+import TeamGearTable from './team-gear-table';
+import TeamInfoTable from './team-info-table';
 const { openExternal } = window.require('electron').remote.shell;
 
 const ResultSummary = ({ result }) =>
@@ -73,7 +86,7 @@ const ResultSummary2 = ({ result }) =>
         : null}
       {result.estimate_gachi_power
         ? <tr>
-            <th>Estimate Power</th>
+            <th>Estimate Gachi Power</th>
             <td>
               {result.estimate_gachi_power}
             </td>
@@ -81,7 +94,7 @@ const ResultSummary2 = ({ result }) =>
         : null}
       {result.fes_power
         ? <tr>
-            <th>Power</th>
+            <th>Fes Power</th>
             <td>
               {result.fes_power}
             </td>
@@ -89,7 +102,7 @@ const ResultSummary2 = ({ result }) =>
         : null}
       {result.my_estimate_league_point
         ? <tr>
-            <th>Estimate League Power</th>
+            <th>My Estimate League Power</th>
             <td>
               {result.my_estimate_league_point}
             </td>
@@ -114,216 +127,115 @@ const ResultSummary2 = ({ result }) =>
     </tbody>
   </Table>;
 
-const PlayerRow = ({ player }) => {
-  return (
-    <tr>
-      <td>
-        {player.player.nickname}
-      </td>
+class ResultDetailCard extends React.Component {
+  state = {
+    show: 1
+  };
 
-      {player.player.udemae
-        ? <td>
-            {`${player.player.udemae.name}${player.player.udemae.s_plus_number
-              ? player.player.udemae.s_plus_number
-              : player.player.udemae.name === 'S+' ? 0 : ''}`}
-          </td>
-        : null}
-      <td>
-        {player.player.weapon.name}
-      </td>
-      <td>
-        {player.game_paint_point}
-      </td>
-      <td>
-        {player.kill_count + player.assist_count}
-      </td>
-      <td>
-        {player.kill_count}
-      </td>
-      <td>
-        {player.assist_count}
-      </td>
-      <td>
-        {player.death_count}
-      </td>
-      <td>
-        {player.special_count}
-      </td>
-    </tr>
-  );
-};
+  showStats = () => {
+    this.setState({ show: 1 });
+  };
 
-const TeamHeader = ({ player = { player: {} } }) =>
-  <thead>
-    <tr>
-      <th>Player</th>
-      {player.player.udemae ? <th>Rank</th> : null}
-      <th>Weapon</th>
-      <th>Inked</th>
-      <th>K+A</th>
-      <th>K</th>
-      <th>A</th>
-      <th>D</th>
-      <th>S</th>
-    </tr>
-  </thead>;
+  showGear = () => {
+    this.setState({ show: 2 });
+  };
 
-const MyTeamTable = ({ result }) => {
-  const total_k =
-    result.my_team_members.reduce((sum, player) => sum + player.kill_count, 0) +
-    result.player_result.kill_count;
-  const total_a =
-    result.my_team_members.reduce(
-      (sum, player) => sum + player.assist_count,
-      0
-    ) + result.player_result.assist_count;
+  showInfo = () => {
+    this.setState({ show: 3 });
+  };
 
-  return (
-    <Table striped bordered condensed hover>
-      <TeamHeader player={result.player_result} />
-      <tbody>
-        <PlayerRow player={result.player_result} />
-        {result.my_team_members.map(player =>
-          <PlayerRow key={player.player.nickname} player={player} />
-        )}
-      </tbody>
-      <tfoot>
-        <tr>
-          <th>Totals</th>
-          <th />
-          {result.player_result.player.udemae ? <th /> : null}
-          <td>
-            {result.my_team_members.reduce(
-              (sum, player) => sum + player.game_paint_point,
-              0
-            ) + result.player_result.game_paint_point}
-          </td>
-          <td>
-            {total_k + total_a}
-          </td>
-          <td>
-            {total_k}
-          </td>
-          <td>
-            {total_a}
-          </td>
-          <td>
-            {result.my_team_members.reduce(
-              (sum, player) => sum + player.death_count,
-              0
-            ) + result.player_result.death_count}
-          </td>
-          <td>
-            {result.my_team_members.reduce(
-              (sum, player) => sum + player.special_count,
-              0
-            ) + result.player_result.special_count}
-          </td>
-        </tr>
-      </tfoot>
-    </Table>
-  );
-};
+  render() {
+    const { result, statInk } = this.props;
+    const linkInfo = statInk[result.battle_number];
+    if (!result) {
+      return null;
+    }
 
-const TheirTeamTable = ({ result }) => {
-  const total_k = result.other_team_members.reduce(
-    (sum, player) => sum + player.kill_count,
-    0
-  );
-  const total_a = result.other_team_members.reduce(
-    (sum, player) => sum + player.assist_count,
-    0
-  );
+    const myTeam = result.my_team_members.slice(0);
+    myTeam.unshift(result.player_result);
+    myTeam.sort((a, b) => b.sort_score - a.sort_score);
+    const otherTeam = result.other_team_members
+      .slice(0)
+      .sort((a, b) => b.sort_score - a.sort_score);
 
-  return (
-    <Table striped bordered condensed hover>
-      <TeamHeader player={result.other_team_members[0]} />
-      <tbody>
-        {result.other_team_members.map(player =>
-          <PlayerRow key={player.player.nickname} player={player} />
-        )}
-      </tbody>
-      <tfoot>
-        <tr>
-          <th>Totals</th>
-          <th />
-          {result.player_result.player.udemae ? <th /> : null}
-          <td>
-            {result.other_team_members.reduce(
-              (sum, player) => sum + player.game_paint_point,
-              0
-            )}
-          </td>
-          <td>
-            {total_k + total_a}
-          </td>
-          <td>
-            {total_k}
-          </td>
-          <td>
-            {total_a}
-          </td>
-          <td>
-            {result.other_team_members.reduce(
-              (sum, player) => sum + player.death_count,
-              0
-            )}
-          </td>
-          <td>
-            {result.other_team_members.reduce(
-              (sum, player) => sum + player.special_count,
-              0
-            )}
-          </td>
-        </tr>
-      </tfoot>
-    </Table>
-  );
-};
-
-const ResultDetailCard = ({ result, statInk }) => {
-  const linkInfo = statInk[result.battle_number];
-  return (
-    <Panel
-      header={
-        <h3>
-          {`Battle #${result.battle_number} Details`}
-          {linkInfo
-            ? <a
-                onClick={() =>
-                  openExternal(
-                    `https://stat.ink/@${linkInfo.username}/spl2/${linkInfo.battle}`
-                  )}
-                style={{ cursor: 'pointer' }}
-              >
-                <Glyphicon glyph={'ok-sign'} style={{ paddingLeft: 6 }} />
-              </a>
-            : null}
-        </h3>
-      }
-    >
-      <Grid fluid>
-        <Row>
-          <Col sm={6} md={6}>
-            <ResultSummary result={result} />
-          </Col>
-          <Col sm={6} md={6}>
-            <ResultSummary2 result={result} />
-          </Col>
-        </Row>
-        <Row>
-          <Col sm={6} md={6}>
-            <h4>My Team</h4>
-            <MyTeamTable result={result} />
-          </Col>
-          <Col sm={6} md={6}>
-            <h4>Enemy Team</h4>
-            <TheirTeamTable result={result} />
-          </Col>
-        </Row>
-      </Grid>
-    </Panel>
-  );
-};
+    return (
+      <Panel
+        header={
+          <h3>
+            {`Battle #${result.battle_number} Details`}
+            {linkInfo
+              ? <a
+                  onClick={() =>
+                    openExternal(
+                      `https://stat.ink/@${linkInfo.username}/spl2/${linkInfo.battle}`
+                    )}
+                  style={{ cursor: 'pointer' }}
+                >
+                  <Glyphicon glyph={'ok-sign'} style={{ paddingLeft: 6 }} />
+                </a>
+              : null}
+          </h3>
+        }
+      >
+        <Grid fluid>
+          <Row>
+            <Col sm={6} md={6}>
+              <ResultSummary result={result} />
+            </Col>
+            <Col sm={6} md={6}>
+              <ResultSummary2 result={result} />
+            </Col>
+          </Row>
+          <Row>
+            <Col md={12}>
+              <ButtonToolbar style={{ marginBottom: '10px' }}>
+                <ButtonGroup>
+                  <Button
+                    onClick={this.showStats}
+                    active={this.state.show === 1}
+                  >
+                    Stats
+                  </Button>
+                  <Button
+                    onClick={this.showGear}
+                    active={this.state.show === 2}
+                  >
+                    Gear
+                  </Button>
+                  <Button
+                    onClick={this.showInfo}
+                    active={this.state.show === 3}
+                  >
+                    More Info
+                  </Button>
+                </ButtonGroup>
+              </ButtonToolbar>
+            </Col>
+          </Row>
+          <Row>
+            <Col sm={6} md={6}>
+              <h4>My Team</h4>
+              {this.state.show === 1 ? <TeamStatsTable team={myTeam} /> : null}
+              {this.state.show === 2 ? <TeamGearTable team={myTeam} /> : null}
+              {this.state.show === 3 ? <TeamInfoTable team={myTeam} /> : null}
+            </Col>
+            <Col sm={6} md={6}>
+              <h4>Enemy Team</h4>
+              {this.state.show === 1
+                ? <TeamStatsTable team={otherTeam} />
+                : null}
+              {this.state.show === 2
+                ? <TeamGearTable team={otherTeam} />
+                : null}
+              {this.state.show === 3
+                ? <TeamInfoTable team={otherTeam} />
+                : null}
+            </Col>
+          </Row>
+        </Grid>
+      </Panel>
+    );
+  }
+}
 
 export default ResultDetailCard;